Fix signed integer overflow in diagram count calculations
The pown() function was causing signed integer overflow when computing
the number of diagrams for OCA and higher-order expansions. With P=98
poles, the intermediate squaring operations in the exponentiation-by-
squaring algorithm overflowed 32-bit integers.

Changes:
- Modified pown() to use int64_t instead of template-based types
- Updated number_of_diagrams() return type to int64_t
- Changed all diagram count variables (total_num_diagram,
  num_diagram_per_fb, etc.) to int64_t throughout impurity.cpp and
  strong_cpl.cpp
- Added <cstdint> include to utils.hpp

This prevents UndefinedBehaviorSanitizer errors and allows the solver
to handle larger pole counts without overflow.

+++ c++/triqs_soehyb/strong_cpl.hpp
@@ -120,15 +120,15 @@ nda::array<dcomplex, 3> Sigma_Diagram_calc(hyb_F &hyb_F_self, hyb_F &hyb_F_refle
                                            nda::array_const_view<dcomplex, 3> Deltat, nda::array_const_view<dcomplex, 3> Deltat_reflect,
                                            nda::array_const_view<dcomplex, 3> Gt, imtime_ops &itops, double beta,
                                            nda::array_const_view<dcomplex, 3> F, nda::array_const_view<dcomplex, 3> F_dag,
-                                           nda::vector_const_view<int> fb, bool backward = true);
+                                           nda::vector_const_view<int64_t> fb, bool backward = true);
 
-/** 
+/**
 * @brief Evaluating impurity Green's function diagram with given topology and forward/backward flag. Input arguments are the same as Sigma_Diagram_calc
 * @return impurity Green's function r*n*n
 * */
 nda::array<dcomplex, 3> G_Diagram_calc(hyb_F &hyb_F_self, hyb_F &hyb_F_reflect, nda::array_const_view<int, 2> D,
                                        nda::array_const_view<dcomplex, 3> Gt, imtime_ops &itops, double beta, nda::array_const_view<dcomplex, 3> F,
-                                       nda::array_const_view<dcomplex, 3> F_dag, nda::vector_const_view<int> fb);
+                                       nda::array_const_view<dcomplex, 3> F_dag, nda::vector_const_view<int64_t> fb);
 
 /** 
 * @brief Evaluating all pseudo-particle self energy diagram with given topology, which sum over all forward and backward choices. For explanation of input parameters, see Sigma_Diagram_calc.
@@ -160,7 +160,8 @@ nda::array<dcomplex, 3> Sigma_OCA_calc(hyb_F &hyb_F, nda::array_const_view<dcomp
 * @return impurity Green's function r*n*n
 * */
 nda::array<dcomplex, 3> G_OCA_calc(hyb_F &hyb_F_self, hyb_F &hyb_F_reflect, nda::array_const_view<dcomplex, 3> Gt, imtime_ops &itops, double beta,
-                                   nda::array_const_view<dcomplex, 3> F, nda::array_const_view<dcomplex, 3> F_dag, nda::vector_const_view<int> fb);
+                                   nda::array_const_view<dcomplex, 3> F, nda::array_const_view<dcomplex, 3> F_dag,
+                                   nda::vector_const_view<int64_t> fb);
 
 /** 
 * @brief Gt(k,_,_) = matmul(Ft(k,_,_), Gt(k,_,_))
@@ -196,11 +197,11 @@ nda::array<dcomplex, 3> evaluate_one_diagram(hyb_F &hyb_F_self, hyb_F &hyb_F_ref
                                              nda::array_const_view<dcomplex, 3> Deltat, nda::array_const_view<dcomplex, 3> Deltat_reflect,
                                              nda::array_const_view<dcomplex, 3> Gt, imtime_ops &itops, double beta,
                                              nda::array_const_view<dcomplex, 3> F, nda::array_const_view<dcomplex, 3> F_dag,
-                                             nda::vector_const_view<int> fb, bool backward, int num0, int m, int n, int r, int N, int P);
+                                             nda::vector_const_view<int64_t> fb, bool backward, int num0, int m, int n, int r, int N, int P);
 
 nda::array<dcomplex, 3> eval_one_diagram_G(hyb_F &hyb_F_self, hyb_F &hyb_F_reflect, nda::array_const_view<int, 2> D,
                                            nda::array_const_view<dcomplex, 3> Gt, nda::array_const_view<dcomplex, 3> Gt_reflect, imtime_ops &itops,
                                            double beta, nda::array_const_view<dcomplex, 3> F, nda::array_const_view<dcomplex, 3> F_dag,
-                                           nda::vector_const_view<int> fb, int num0, int m, int n, int r, int N, int P);
+                                           nda::vector_const_view<int64_t> fb, int num0, int m, int n, int r, int N, int P);
 
 dcomplex trace_matmul(nda::array_const_view<dcomplex, 2> M1, nda::array_const_view<dcomplex, 2> M2);
